[change] Minor changes
