+++ README.rst
@@ -411,7 +411,7 @@ To register a signal you need to include following code in your ``apps.py``.
 **Note**: You need to import ``register_notification_cache_update`` inside the ``ready`` function or 
 you can define another funtion to register signals which will be called in ``ready`` and then it will be 
 imported in this function. Also ``dispatch_uid`` is unique identifier of a signal. You can pass any 
-value you want but it needs to be unique. For more details read `preventing-duplicate-signals <https://docs.djangoproject.com/en/dev/topics/signals/#preventing-duplicate-signals>`_
+value you want but it needs to be unique. For more details read `preventing duplicate signals section of Django documentation <https://docs.djangoproject.com/en/dev/topics/signals/#preventing-duplicate-signals>`_
 
 Notification Types
 ------------------
